remove contact form and fix links
this form currently does not work, the send.php script is not executed,
if accessed directly, it gets downloaded/displayed in browser. also as a
spam prevention. the removed javascript files were used only in the
contact form, so safe to remove.

this commit also fixes some links that were either broken or were not
using compilation variables (e.g. for Lakka github page).